refactor: simplify data integrity issues rendering
