+++ src/pages/data-integrity/Issues/IssueCard.module.css
@@ -2,6 +2,10 @@
     margin-bottom: 10px;
 }
 
+.cardText {
+    padding-top: 0 !important;
+}
+
 .card h4 {
     margin-top: 0;
     margin-bottom: 0;
