+++ src/pages/data-integrity/Issues/Issues.module.css
@@ -1,0 +1,3 @@
+.issues {
+    margin-bottom: var(--spacers-dp16);
+}
